replica_rac2: move RaftNode from kvserver
Epic: none
Release note: none
