+++ pkg/kv/kvserver/kvflowcontrol/replica_rac2/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
         "close_scheduler.go",
         "doc.go",
         "processor.go",
+        "raft_node.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/kv/kvserver/kvflowcontrol/replica_rac2",
     visibility = ["//visibility:public"],
@@ -15,7 +16,9 @@ go_library(
         "//pkg/kv/kvserver/kvflowcontrol/kvflowcontrolpb",
         "//pkg/kv/kvserver/kvflowcontrol/rac2",
         "//pkg/kv/kvserver/raftlog",
+        "//pkg/raft",
         "//pkg/raft/raftpb",
+        "//pkg/raft/tracker",
         "//pkg/roachpb",
         "//pkg/settings/cluster",
         "//pkg/util/admission/admissionpb",
